Flush compressed data out of encoders more often (#383)
* Flush compressed data out of encoders more often

Currently flush() is not getting called so compressed output is not
returned to the client the encoder until input is fully read(). In my use case,
flushing more often is necessary as otherwise it appears to the client
the data gets 'buffered'.

There's been a previous PR here and some discussion
https://github.com/Nullus157/async-compression/pull/155
but appears to be abandonded so if there's interest in this fix, this PR
tries this again.
The tests are passing with this change.

* Update futures bufread futures encoder to add more flushes

It's equivalent to the tokio implementation.